After moving controllers to their respective store modules, clean up sessions tests
