+++ authn/webapi/store/roles/controller/controller_test.go
@@ -1,7 +1,7 @@
 // brian taylor vann
 // taylorvann dot com
 
-package roles
+package controller
 
 import (
 	"testing"
